requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
websockets==12.0
pydantic==2.5.2
loguru==0.7.2
fastapi==0.104.1
uvicorn==0.24.0